Added rust benchmark
Note that rust benchmark uses not only rust but cargo as well.
An original git is found here: https://github.com/White-Oak/rust-simple-bench , It consists of two solutions: on-stack and in-heap.
Both are run in a considerable time, being slighter faster than Java.
This commit adds an in-heap solution to a benchmark as it is found to be a little bit quicker than another one during a local testing.
Also it looks prettier than another one :)

Code writes an integer from the calculated array to /dev/null, so the code isn't removed by an optimizing compiler when using --release
option.